+++ onlyoffice_odoo/__manifest__.py
@@ -29,6 +29,7 @@
             "onlyoffice_odoo/static/src/models/*.js",
         ],
         "web.assets_backend": [
+            "onlyoffice_odoo/static/src/actions/*",
             "onlyoffice_odoo/static/src/components/*/*.xml",
         ],
     },